refactor: Switch from prepending to conditions
